pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arche-browser"
version = "2.2.0"
description = "MCP Server for Browser Automation and Local PC Control"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "GizAI"}
]
keywords = [
    "mcp", "browser", "automation", "chrome", "cdp",
    "devtools", "claude", "selenium", "puppeteer", "playwright"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]
dependencies = [
    "mcp>=1.0.0",
    "websocket-client>=1.0.0",
    "requests>=2.20.0",
    "uvicorn>=0.20.0",
    "starlette>=0.27.0",
    "cryptography>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "mypy", "ruff"]

[project.scripts]
arche-browser = "arche_browser.__main__:main"

[project.urls]
Homepage = "https://github.com/GizAI/arche-browser"
Repository = "https://github.com/GizAI/arche-browser"
Issues = "https://github.com/GizAI/arche-browser/issues"

[tool.setuptools.packages.find]
include = ["arche_browser*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
